translation_tool_auto/requirements.txt
streamlit
scikit-learn
numpy
requests
python-dotenv
